helpModal util
