+++ lib/ui/elements/_elements.mjs
@@ -15,6 +15,7 @@ import dropdown_multi from './dropdown_multi.mjs';
 import btnPanel from './btnPanel.mjs';
 import legendIcon from './legendIcon.mjs';
 import modal from './modal.mjs';
+import helpModal from './helpModal.mjs';
 import searchbox from './searchbox.mjs';
 import slider from './slider.mjs';
 import slider_ab from './slider_ab.mjs';
@@ -55,6 +56,7 @@ export default {
   dropdown_multi,
   legendIcon,
   modal,
+  helpModal,
   searchbox,
   slider,
   slider_ab,
