+++ lib/ui/elements/drawing.mjs
@@ -160,34 +160,6 @@ mapp.utils.merge(mapp.dictionaries, {
   }
 })
 
-function helpModal(content) {
-
-  // If a modal is already present, remove it
-  const modal = document.querySelector('[data-id=modal_drawing]')
-  
-  modal && modal.remove()
-
-  if (!content) return;
-
-  // Add the header
-  const header = mapp.utils.html`<h3>${mapp.dictionary.draw_modal_title}</h3>`;
-
-  // Create the modal
-  mapp.ui.elements.modal({
-    data_id: 'modal_drawing',
-    header: header,
-    target: document.getElementById('Map'),
-    content: content,
-    height: 'auto',
-    width: '200px',
-    css_style: 'padding: 0.5em;',
-    top: 30,
-    left: 60,
-    contained: true,
-    close: true
-  });
-}
-
 function drawOnclick(e, layer, interaction) {
 
   const btn = e.target
@@ -208,7 +180,7 @@ function drawOnclick(e, layer, interaction) {
 
     delete layer.mapview.interaction
 
-    helpModal();
+    mapp.ui.elements.helpModal();
 
     // Set highlight interaction if no other interaction is current after 400ms.
     setTimeout(() => {
@@ -218,7 +190,10 @@ function drawOnclick(e, layer, interaction) {
 
   layer.mapview.interactions.draw(interaction)
 
-  helpModal(interaction.helpModal);
+  // Add the header
+  interaction.helpModal.header = mapp.utils.html`<h3>${mapp.dictionary.draw_modal_title}</h3>`;
+
+  mapp.ui.elements.helpModal(interaction.helpModal);
 
   btn.classList.add('active')
 }
@@ -232,9 +207,11 @@ function point(layer) {
   }, typeof layer.draw.point === 'object' && layer.draw.point || {})
 
 
-  layer.draw.point.helpModal ??= mapp.utils.html.node`<li>
+  layer.draw.point.helpModal ??= {
+    content: mapp.utils.html.node`<li>
     <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
     <ul>${mapp.dictionary.draw_modal_save_single}</ul>`
+  }
 
   // If a label is provided, use it, otherwise use the default
   let label = layer.draw.point?.label || mapp.dictionary.draw_point
@@ -258,10 +235,12 @@ function line(layer) {
     type: 'LineString',
   }, typeof layer.draw.line === 'object' && layer.draw.line || {})
 
-  layer.draw.line.helpModal ??= mapp.utils.html.node`<li>
-  <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
-  <ul>${mapp.dictionary.draw_modal_remove_vertex}</ul>
-  <ul>${mapp.dictionary.draw_modal_save_single}</ul>`
+  layer.draw.line.helpModal ??= {
+    content: mapp.utils.html.node`<li>
+    <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
+    <ul>${mapp.dictionary.draw_modal_remove_vertex}</ul>
+    <ul>${mapp.dictionary.draw_modal_save_single}</ul>`
+  }
 
   // If a label is provided, use it, otherwise use the default
   let label = layer.draw.line?.label || mapp.dictionary.draw_line
@@ -287,11 +266,13 @@ function polygon(layer) {
   // If a label is provided, use it, otherwise use the default
   let label = layer.draw.polygon?.label || mapp.dictionary.draw_polygon
 
-  layer.draw.polygon.helpModal ??= mapp.utils.html.node`<li>
+  layer.draw.polygon.helpModal ??= {
+    content: mapp.utils.html.node`<li>
     <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
     <ul>${mapp.dictionary.draw_modal_cancel_drawing}
     <ul>${mapp.dictionary.draw_modal_remove_vertex}</ul>
     <ul>${mapp.dictionary.draw_modal_save}</ul>`
+  }
 
   // Create the button
   layer.draw.polygon.btn = mapp.utils.html.node`
@@ -312,9 +293,11 @@ function rectangle(layer) {
     geometryFunction: ol.interaction.Draw.createBox(),
   }, typeof layer.draw.rectangle === 'object' && layer.draw.rectangle || {})
 
-  layer.draw.rectangle.helpModal ??= mapp.utils.html.node`<li>
+  layer.draw.rectangle.helpModal ??= {
+    content: mapp.utils.html.node`<li>
     <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
     <ul>${mapp.dictionary.draw_modal_save_single}</ul>`
+  }
 
   // If a label is provided, use it, otherwise use the default
   let label = layer.draw.rectangle?.label || mapp.dictionary.draw_rectangle
@@ -340,9 +323,11 @@ function circle_2pt(layer) {
     ...layer.draw.circle_2pt
   }
 
-  layer.draw.circle_2pt.helpModal ??= mapp.utils.html.node`<li>
+  layer.draw.circle_2pt.helpModal ??= {
+    content: mapp.utils.html.node`<li>
     <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
     <ul>${mapp.dictionary.draw_modal_save_single}</ul>`
+  }
 
   // Create the button
   layer.draw.circle_2pt.btn = mapp.utils.html.node`
@@ -387,9 +372,11 @@ function circle(layer) {
     ...layer.draw.circle
   }
 
-  layer.draw.circle.helpModal ??= mapp.utils.html.node`<li>
+  layer.draw.circle.helpModal ??= {
+    content: mapp.utils.html.node`<li>
     <ul>${mapp.dictionary.draw_modal_begin_drawing}</ul>
     <ul>${mapp.dictionary.draw_modal_save_single}</ul>`
+  }
 
   const unitsDropDown = mapp.utils.html.node`
     <div style="display: grid; grid-template-columns: 100px 1fr; align-items: center;">
